highp to mediump

+++ app/src/shaders/es/font.frag
@@ -1,5 +1,5 @@
 #version 300 es
-precision highp float;
+precision mediump float;
 
 in vec2 TexCoords;
 out vec4 color;
